% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.batch_operations.R
\name{delete_compute_environment}
\alias{delete_compute_environment}
\title{Deletes an AWS Batch compute environment}
\usage{
delete_compute_environment(computeEnvironment)
}
\arguments{
\item{computeEnvironment}{[required] The name or Amazon Resource Name (ARN) of the compute environment to delete.}
}
\description{
Deletes an AWS Batch compute environment.
}
\details{
Before you can delete a compute environment, you must set its state to \code{DISABLED} with the UpdateComputeEnvironment API operation and disassociate it from any job queues with the UpdateJobQueue API operation.
}
\section{Accepted Parameters}{
\preformatted{delete_compute_environment(
  computeEnvironment = "string"
)
}
}

\examples{
# This example deletes the P2OnDemand compute environment.
\donttest{delete_compute_environment(
  computeEnvironment = "P2OnDemand"
)}

}
