% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.batch_operations.R
\name{delete_job_queue}
\alias{delete_job_queue}
\title{Deletes the specified job queue}
\usage{
delete_job_queue(jobQueue)
}
\arguments{
\item{jobQueue}{The short name or full Amazon Resource Name (ARN) of the queue to delete.}
}
\description{
Deletes the specified job queue. You must first disable submissions for a queue with the UpdateJobQueue operation. All jobs in the queue are terminated when you delete a job queue.
}
\details{
It is not necessary to disassociate compute environments from a queue before submitting a \code{DeleteJobQueue} request.
}
\examples{
# This example deletes the GPGPU job queue.
delete_job_queue(
  jobQueue = "GPGPU"
)

}
