% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.batch_operations.R
\name{deregister_job_definition}
\alias{deregister_job_definition}
\title{Deregisters an AWS Batch job definition}
\usage{
deregister_job_definition(jobDefinition)
}
\arguments{
\item{jobDefinition}{[required] The name and revision (\code{name:revision}) or full Amazon Resource Name (ARN) of the job definition to deregister.}
}
\description{
Deregisters an AWS Batch job definition.
}
\section{Accepted Parameters}{
\preformatted{deregister_job_definition(
  jobDefinition = "string"
)
}
}

\examples{
# This example deregisters a job definition called sleep10.
\donttest{deregister_job_definition(
  jobDefinition = "sleep10"
)}

}
