% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.batch_operations.R
\name{describe_jobs}
\alias{describe_jobs}
\title{Describes a list of AWS Batch jobs}
\usage{
describe_jobs(jobs)
}
\arguments{
\item{jobs}{A space-separated list of up to 100 job IDs.}
}
\description{
Describes a list of AWS Batch jobs.
}
\examples{
# This example describes a job with the specified job ID.
describe_jobs(
  jobs = list(
    "24fa2d7a-64c4-49d2-8b47-f8da4fbde8e9"
  )
)

}
