% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.batch_package.R
\docType{package}
\name{paws.batch-package}
\alias{paws.batch}
\alias{paws.batch-package}
\title{paws.batch: AWS Batch}
\description{
AWS Batch enables you to run batch computing workloads on the
AWS Cloud. Batch computing is a common way for developers, scientists,
and engineers to access large amounts of compute resources, and AWS
Batch removes the undifferentiated heavy lifting of configuring and
managing the required infrastructure. AWS Batch will be familiar to
users of traditional batch computing software. This service can
efficiently provision resources in response to jobs submitted in order
to eliminate capacity constraints, reduce compute costs, and deliver
results quickly.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
