% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.batch_operations.R
\name{update_compute_environment}
\alias{update_compute_environment}
\title{Updates an AWS Batch compute environment}
\usage{
update_compute_environment(
  computeEnvironment = "string",
  state = "ENABLED"|"DISABLED",
  computeResources = list(
    minvCpus = 123,
    maxvCpus = 123,
    desiredvCpus = 123
  ),
  serviceRole = "string"
)
}
\arguments{
\item{computeEnvironment}{[required] The name or full Amazon Resource Name (ARN) of the compute environment to update.}

\item{state}{The state of the compute environment. Compute environments in the \code{ENABLED} state can accept jobs from a queue and scale in or out automatically based on the workload demand of its associated queues.}

\item{computeResources}{Details of the compute resources managed by the compute environment. Required for a managed compute environment.}

\item{serviceRole}{The full Amazon Resource Name (ARN) of the IAM role that allows AWS Batch to make calls to other AWS services on your behalf.

If your specified role has a path other than \code{/}, then you must either specify the full role ARN (this is recommended) or prefix the role name with the path.

Depending on how you created your AWS Batch service role, its ARN may contain the \code{service-role} path prefix. When you only specify the name of the service role, AWS Batch assumes that your ARN does not use the \code{service-role} path prefix. Because of this, we recommend that you specify the full ARN of your service role when you create compute environments.}
}
\description{
Updates an AWS Batch compute environment.
}
\examples{
# This example disables the P2OnDemand compute environment so it can be
# deleted.
update_compute_environment(
  computeEnvironment = "P2OnDemand",
  state = "DISABLED"
)

}
